fixed CUDA kernel bounds error
If you ever did anything with a vector of size greater than 4096, the tail of the vector literally was not operated on at all by the kernel. I had copied the two constants from a cuLaunchKernel() example and left a TODO to check what they meant later. Turns out, they basically controlled the size of the input vector.