added TODO
NVRTC is truly evil. It does weird things with interrupt handlers. On Linux, we had to prevent it from changing a stack-related flag so that Go could capture interrupts. On OS X, it seems to cause SIGINTs to trigger segmentation faults. It is fairly easy to avoid NVRTC: we just need to compile the PTX before-hand. I will likely create a Makefile for doing this, that way the user doesn't have to do it unless they want to change the kernels themselves.