cuda: custom GC monitoring
Basically, the go GC isn't very aggressive when it doesn't think it's using very much memory. And yes, that is a good thing, but it allows the CUDA memory to grow out of control and eventually crash the program. Now, using a special GC mechanism, we can force the runtime to collect unused buffers once the memory starts growing too much.