make some operations async
This has one huge advantage: it helps to avoid the large overhead of switching back and forth between a pinned Goroutine. Whether or not it helps in practice is another story.

Also, this is definitely not ready for the master branch. First, I'll want to clean up the argument scheme for ops32 in general.