introduced allocator type & handle Close() method
I realized that a finalizer on Handle is a bad idea. Suppose a call is using a resource (e.g. the math kernels). That function might not have a reference to the handle itself (but only to the math kernels), so the handle's finalizer might be called, freeing the math kernels and causing bad things.